### 2021-02-12 #3033 @whedon
[PRE REVIEW]:  toqito -- Theory of quantum information toolkit: A Python package for studying quantum information
**Submitting author:** @vprusso (<a href="http://orcid.org/0000-0002-8952-8981">Vincent Russo</a>)
**Repository:** <a href="https://github.com/vprusso/toqito" target ="_blank">https://github.com/vprusso/toqito</a>
**Version:** 0.0.7
