### 2021-02-12 #3033 @vprusso
Hi Afron,

Thanks again for your quick response.

I've gone ahead and added in a "Statement of Need" section along with the
"doi" information in the BibTeX file for the entries that do possess a
valid doi.

If there is anything else I need to add, please let me know, and thanks
again.

Vincent

On Fri, Feb 12, 2021 at 4:11 AM Arfon Smith <notifications@github.com>
wrote:

> @vprusso <https://github.com/vprusso> - please update your paper to
> include a section named Statement of need with a description that
> illustrates the research purpose of the software. Also, I'm not sure if
> those DOI recommendations from Whedon are correct but if they are, please
> add them doi fields to your BibTeX.
>
