### 2021-02-20 #3033 @Kevin-Mattheus-Moerman
@vprusso see :point_up:, can you remove the `https://doi.org/` prefix to fix these references?
