### 2021-02-20 2c75209
Fixing bibtex references
