+++ paper/paper.bib
@@ -13,7 +13,7 @@
   title={{QETLAB}: A {MATLAB} toolbox for quantum entanglement, version 0.9},
   author={Johnston, Nathaniel},
   journal={qetlab.com},
-  doi={https://doi.org/10.5281/zenodo.44637},
+  doi={10.5281/zenodo.44637},
   year={2016}
 }
 
@@ -25,7 +25,7 @@
   number={4},
   pages={1234--1240},
   year={2013},
-  doi={https://doi.org/10.1016/j.cpc.2012.11.019},
+  doi={10.1016/j.cpc.2012.11.019},
   publisher={Elsevier}
 }
 
@@ -36,7 +36,7 @@
   volume={3},
   pages={129},
   year={2019},
-  doi={https://doi.org/10.22331/q-2019-03-11-129},
+  doi={10.22331/q-2019-03-11-129},
   publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften}
 }
 
@@ -46,7 +46,7 @@
   journal={Quantum},
   volume={2},
   pages={49},
-  doi={https://doi.org/10.22331/q-2018-01-31-49},
+  doi={10.22331/q-2018-01-31-49},
   year={2018},
   publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften}
 }
@@ -55,7 +55,7 @@
   title={Quantum++--A C++ 11 quantum computing library},
   author={Gheorghiu, Vlad},
   journal={arXiv preprint arXiv:1412.4704},
-  doi={https://doi.org/10.1371/journal.pone.0208073},
+  doi={10.1371/journal.pone.0208073},
   year={2014}
 }
 
