### 2021-02-20 #3033 @vprusso
Fixed the `https://doi.org/` prefix in the references.

On Sat, Feb 20, 2021 at 6:29 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @jedbrown <https://github.com/jedbrown> @poulson
> <https://github.com/poulson> @gkthiruvathukal
> <https://github.com/gkthiruvathukal> could one of you handle this
> submission? Thanks
>
