### 2021-02-22 #3033 @poulson
Hi @Kevin-Mattheus-Moerman -- I am pretty comfortable with the convex optimization and linear algebra they are making use of but am pretty far from an expert on quantum information theory. I should be able to handle finding reviewers who are though.
