### 2021-02-22 #3033 @poulson
Hi Vincent (@vprusso) :wave:.

Thank you for this submission and, before I reach out to find a set of reviewers, are there any experts in both quantum information theory (and, ideally, python) that you would recommend?
