### 2021-02-23 #3033 @vprusso
Hi Jack,

Thanks for getting back to me. Sure, I can suggest a few individuals who
are quantum information theorists who have also contributed to various
Python projects that I think would be great candidates (assuming they are
available, of course):

Sarah Kaiser: https://github.com/crazy4pi314
Josh Combes: https://github.com/joshcombes
Ryan LaRose: https://github.com/rmlarose
Nathan Shammah: https://github.com/nathanshammah
Andrea Mari: https://github.com/andreamari

Let me know what you think, and thanks again.

Vincent

On Mon, Feb 22, 2021 at 7:58 AM Jack Poulson <notifications@github.com>
wrote:

> Hi Vincent (@vprusso <https://github.com/vprusso>) 👋.
>
> Thank you for this submission and, before I reach out to find a set of
> reviewers, are there any experts in both quantum information theory (and,
> ideally, python) that you would recommend?
>
