### 2021-02-23 #3033 @gkthiruvathukal
@vprusso I was probably not clear on this request. Are they currently on the JOSS list of reviewers? (See link at the very top of this issue thread). It would be ideal if you can select from that list. I am working to clarify how new reviewers can be added to the system.
