### 2021-02-23 #3033 @vprusso
Ah, my apologies. Sorry about that. Here are some JOSS reviewers who I
think would be well-suited for the submission:

jameshclrk
goerz
nmoran
ycchen1989
TejasAvinashShetty
WhiteSymmetry
marwahaha
rurz

Let me know if I can provide anything further, and thanks again.

On Mon, Feb 22, 2021 at 10:06 PM George K. Thiruvathukal <
notifications@github.com> wrote:

> @vprusso <https://github.com/vprusso> I was probably not clear on this
> request. Are they currently on the JOSS list of reviewers? (See link at the
> very top of this issue thread). It would be ideal if you can select from
> that list. I am working to clarify how new reviewers can be added to the
> system.
>
