### 2021-02-26 #3033 @poulson
Hi Alejandra (@rurz) -- your colleague Vincent Russo (@vprusso) has recommended you as an expert on quantum information theory that would be an ideal reviewer for their python package in said area.

Please let me know if you would be willing to serve as a reviewer for this Journal of Open Source Software submission.
