### 2021-03-03 #3033 @jameshclrk
> Hi James (@jameshclrk) -- you were recommended by Vincent Russo (@vprusso) as an expert on quantum information theory that would be appropriate for reviewing their submission of a python toolkit in that area.
>
> Would you be willing to serve as one of the reviewers of this submission?

Sorry for the delay in replaying - email notifications weren't working. If you're still looking for a reviewer, I'm happy to help.
