### 2021-03-17 vprusso/toqito#47 @marwahaha
JOSS review comments
from https://github.com/openjournals/joss-reviews/issues/3082

Overall, neat package. Comprehensive tests. I like the docs page and the tutorials. Few small comments:
* contributing.md specifies that "multiple core developers will approve PRs", but there seems to be a single core developer.
* your codecov check fails often (because the codecov decreases by a tiny fraction like 0.05%) which makes it seems like "the checks don't pass". Could that be relaxed to only fail when codecov decreases by 0.5% or similar?
* in the paper, how does your software relate to quantum circuit libraries like qiskit? are the softwares competitive? complementary?
* i'm not sure what the proper citation syntax is -- in the paper, consider including a space between the word and the citation, e.g. "physics(citation)" --> "physics (citation)"?
* in the introductory tutorial, the channels and measurements are not well documented : https://toqito.readthedocs.io/en/latest/intro_tutorial.html
* typo in channel docs: depolarizng -> depolarizing
* . How is the randomness generated in toqito.random ?
* what is a perfect matching? the docs claim it takes 2 arguments, but it takes 1 when running the code. i only see outputs for powers of 2.
* the nonlocal game section has a part that looks like this in the docs
![image](https://user-images.githubusercontent.com/2541209/111401707-0a830e00-8687-11eb-833c-53535a9a119a.png)
* are there invalid values for the prob_mat and pred_mat in the nonlocal games? What do the quantum_value and classical_value methods do in this case? (more specifically: how do you know you are using toqito's nonlocal game functionality properly?)

I tried the nonlocal game tutorial and received this error when trying to run the nonlocal game. The xor game was ok.
```
>>> chsh = NonlocalGame(dim, prob_mat, pred_mat)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/kunal/miniconda3/lib/python3.8/site-packages/toqito/nonlocal_games/nonlocal_game.py", line 42, in __init__
    if reps == 1:
ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()
```
* if you think the project is ready, you could 1.0 it, so software users know that they can rely on a consistent interface. I guess this is controversial, see [here](https://neilb.org/2017/04/11/version-one-dot-oh.html) and [here](https://github.com/semver/semver/issues/221), but I would recommend a 1.0.0 release so additional changes are explicit to current users.

Nice work!
