### 2021-03-18 vprusso/toqito#48 @rurz
JOSS Review from rurz
Review for vprusso/toqito

In general, it is an remarkable package, with a lot of resources ranged from basic lineal algebra methods for quantum mechanics, to the implementation of a plethora of quantum measurements, and the development of nonlocal games studies. The package denotes a lot of work, constant planning and implementation of features, as seen in the issues raised on their Github repo.

As stated in the paper, the package intended for quantum information audience, so, no much physical exploration is given across the methods and functions, instead, their logic and mathematical features are polished.

Nice powerful features I found in the package:

* A complete implementation of the basic and advanced mathematical methods to doing basic quantum mechanics developments and checks

* As far as I know, there's not packages who provide such quantities of options to quantum states properties and measurements, some are really useful.

* The package promotions as "quantum information related computer science," and this is clear that we have a work intended more for CS, but in my own view, the package is enough powerful to carry on works on quantum optics.

* The documentation is really great and complete. The presence of a dedicated webpage for the package is a plus.

Some minor issues of my own usage of the package:

* I think the summary of the paper can be improved, since the scope of JOSS is in part provide resources for the non-specialist who seek to known the package. Maybe extend what are the _nonlocal games_ on an extra line.

* Minor typos in the documentation (It denotes that was a human who do the work)
* In the paper a typo: tranpose -> transpose

* I had some problems with the use of `cvxpy` library at first start, this because it was not installed on my environment, and, it seems, it requires a carefully procedure, first _lapack_ and _blas_ libraries, then _cvxpy_. I don't know if this happen to everyone, but maybe a clarification about this could be great, since `cvxpy` is at the core of the package.
* Running `pytest cov` returns a coverage of nearly 95%, this traduces on 56 fails.
* I receive errors from the _Lower bounding the quantum value of the CHSH game_ example tutorial

![Screenshot_2021-03-17_19-13-26](https://user-images.githubusercontent.com/18334633/111562194-615a1780-875b-11eb-91a9-1661db9e8584.png)

* This last error repeats in the _FFL game_. How this errors can be solved to proceed to check what the tutorial claims?
* I have a little knowledge about convex optimization but, what is the justification to construct your package around it? It is more powerful against other procedures? Less resources demanding? Precise?
