### 2021-03-18 #3082 @rurz
Hi, everyone. Hope you're doing well.

I raised an issue thread on the toqito repo for my review -> https://github.com/vprusso/toqito/issues/48

Keep well, and safe
