+++ toqito/perms/perfect_matchings.py
@@ -8,6 +8,9 @@ def perfect_matchings(num: Union[List[int], int, np.ndarray]) -> np.ndarray:
     r"""
     Give all perfect matchings of :code:`num` objects.
 
+    The input can be either an even natural number (the number of objects to be matched) or a
+    `numpy` array containing an even number of distinct objects to be matched.
+
     Returns all perfect matchings of a given list of objects. That is, it returns all ways of
     grouping an even number of objects into pairs.
 
