+++ .codecov.yml
@@ -1,0 +1,10 @@
+# Coverage configuration
+# ----------------------
+coverage:
+  status:
+    patch: false
+
+  range: 70..90     # First number represents red, and second represents green
+  # (default is 70..100)
+  round: down       # up, down, or nearest
+  precision: 2      # Number of decimal places, between 0 and 5
