### 2021-03-20 vprusso/toqito#47 @vprusso
Hi @marwahaha,

Thank you for your outstanding review. Your comments are wonderful, and I really appreciate you taking the time to give this package an honest look and review.

I've addressed your comments in: https://github.com/vprusso/toqito/pull/49.

> contributing.md specifies that "multiple core developers will approve PRs", but there seems to be a single core developer.

Amended to specify myself as the core developer.

> your codecov check fails often (because the codecov decreases by a tiny fraction like 0.05%) which makes it seems like "the checks don't pass". Could that be relaxed to only fail when codecov decreases by 0.5% or similar?

Added a `.codecov.yml` file to (hopefully) resolve that issue.

> i'm not sure what the proper citation syntax is -- in the paper, consider including a space between the word and the citation, e.g. "physics(citation)" --> "physics (citation)"?

I will err on the side of the latter and make the suggested "physics (citation)" convention.

> in the introductory tutorial, the channels and measurements are not well documented : https://toqito.readthedocs.io/en/latest/intro_tutorial.html

There are still quite a few areas to cover, but in the meantime I've placed a "(Coming soon)." marker to indicate that more documentation on a specific subarray will be created there soon.

> typo in channel docs: depolarizng -> depolarizing

Nice catch! Fixed.

> . How is the randomness generated in toqito.random ?

Depending on the random object being generated, this can vary. The overall randomness comes from the `numpy` `random` module. But, for instance, in the case of generated random density matrices, these matrices are generated uniformly at random according to either the Haar or Bures distributions (depending on what the user wants).

> what is a perfect matching? the docs claim it takes 2 arguments, but it takes 1 when running the code. i only see outputs for powers of 2.

I attempted to add a bit more to the docs for this function to clear up this point, as it is a good one.

> the nonlocal game section has a part that looks like this in the docs

Yet again, a nice catch. Fixed!

> are there invalid values for the prob_mat and pred_mat in the nonlocal games? What do the quantum_value and classical_value methods do in this case? (more specifically: how do you know you are using toqito's nonlocal game functionality properly?)

Yes, there are potentially invalid values. Some (basic) checks are in place to make sure that these objects are specified correctly.

> I tried the nonlocal game tutorial and received this error when trying to run the nonlocal game. The xor game was ok.

Looks like my tutorial page was not up-to-date with some of the changes I made to the function. I made sure to update that on the tutorials page.

> if you think the project is ready, you could 1.0 it, so software users know that they can rely on a consistent interface. I guess this is controversial, see here and here, but I would recommend a 1.0.0 release so additional changes are explicit to current users.

That's a good point. I took you up on your advice and went for 1.0.0. Thanks for the suggestion!
