### 2021-03-23 vprusso/toqito#47 @vprusso
Hi @marwahaha. Apologies if I'm interpreting the review process incorrectly, but do I provide my responses and adaptions via a pull request, as I did, and then merge the PR assuming the comments and response are satisfactory? I wanted to ensure I've addressed your comments prior to moving onto the second review. Thanks again for all of your helpful feedback!
