### 2021-03-24 vprusso/toqito#47 @marwahaha
Hi @vprusso ! I don't know what the "best process" is. I did a 30-second glance at the PR and it looks ok, but I'm on vacation and won't be able to really look until next week. I'm ok with you merging it so you can move on to the next review, and I can take another glance next week. Does that work for you?
