### 2021-03-31 vprusso/toqito#47 @marwahaha
A few quick comments:
* contributing.md: I think it should be "approves the reviews" vs "will approve the reviews"
* perfect matchings # of arguments -- maybe it's an issue with the readthedocs generation, but it doesn't look like it's reading the signature correctly:
![image](https://user-images.githubusercontent.com/2541209/113097416-b2cac380-91ab-11eb-934d-69441eba9cdf.png)
* nonlocal games -- I think the tutorial should say `chsh.quantum_value_lower_bound()` not `chsh.quantum_lower_bound()`
* I think the FFL game tutorial has the same issue. Also, the number doesn't match up (I get 0.2222 not 0.6666).
Make sure you push the 1.0 to pypi when you are ready.

Nice work!
