### 2021-04-02 adc81af
Updates from JOSS review
