+++ docs/conf.py
@@ -44,6 +44,7 @@ extensions = ["sphinx.ext.autodoc",
 
 master_doc = "index"
 autosummary_generate = True
+autodoc_typehints = 'none'
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
