### 2021-04-02 vprusso/toqito#47 @vprusso
Hi @marwahaha

> contributing.md: I think it should be "approves the reviews" vs "will approve the reviews"

Agreed, I've gone ahead and made this change.

> perfect matchings # of arguments -- maybe it's an issue with the readthedocs generation, but it doesn't look like it's reading the signature correctly:

Ah, this actually appeared to be some type of subtle problem with `autosummary` (https://github.com/sphinx-doc/sphinx/issues/6311). The fix for me was to apply the feature created in (https://github.com/sphinx-doc/sphinx/issues/6361) and add the line `autodoc_typehints = 'none'` to the `conf.py` file responsible for generating the Sphinx docs. After regenerating with that option, it seems to fix the issue. Nice catch!

> nonlocal games -- I think the tutorial should say chsh.quantum_value_lower_bound() not chsh.quantum_lower_bound()

For this one, I may suggest sticking to `quantum_value_lower_bound()` over `quantum_lower_bound()`. The reason for this is consistency in naming with the other values for nonlocal games--for instance `nonsignaling_value` and `classical_value`. You can let me know if you agree or disagree with this naming logic.

> I think the FFL game tutorial has the same issue. Also, the number doesn't match up (I get 0.2222 not 0.6666).

Nice catch. Thankfully this was just an issue with the tutorial example and not with the code itself. I made sure to test the example and update it as well. Again, great catch!

> Make sure you push the 1.0 to pypi when you are ready.

Awesome, done! :) : https://pypi.org/project/toqito/

> You may also consider giving a talk here: https://unitary.fund/talks.html

I would love to! I've done some prior work with the UnitaryFund and look forward to continuing to do so! Thanks for posting that and keeping the toqito project in mind!
