### 2021-04-02 vprusso/toqito#47 @marwahaha
I think the naming logic is fine. Ctrl-F on https://toqito.readthedocs.io/en/latest/tutorials.nonlocal_games.html for  `quantum_lower_bound` and I think it should instead be `quantum_value_lower_bound`.
