### 2021-04-05 e7ab661
Altering quantum_lower_bound() to quantum_value_lower_bound() in nonlocal game tutorial.
