+++ toqito/matrix_props/is_density.py
@@ -64,7 +64,7 @@ def is_density(mat: np.ndarray) -> bool:
         https://en.wikipedia.org/wiki/Density_matrix
 
     :param mat: Matrix to check.
-    :return: Return `True` if matrix is a density matrix, and `False`
-             otherwise.
+    :return: Return :code:`True` if matrix is a density matrix,
+             and :code:`False` otherwise.
     """
     return is_positive_semidefinite(mat) and np.isclose(np.trace(mat), 1)
