+++ toqito/matrices/fourier.py
@@ -6,8 +6,8 @@ def fourier(dim: int) -> np.ndarray:
     r"""
     Generate the Fourier transform matrix [WikDFT]_.
 
-    Generates the `dim`-by-`dim` unitary matrix that implements the quantum
-    Fourier transform.
+    Generates the :code:`dim`-by-:code:`dim` unitary matrix that implements the
+    quantum Fourier transform.
 
     The Fourier matrix is defined as:
 
@@ -49,7 +49,7 @@ def fourier(dim: int) -> np.ndarray:
         https://en.wikipedia.org/wiki/DFT_matrix
 
     :param dim: The size of the Fourier matrix.
-    :return: The Fourier matrix of dimension `dim`.
+    :return: The Fourier matrix of dimension :code:`dim`.
     """
     # Primitive root of unity.
     root_unity = np.exp(2 * 1j * np.pi / dim)
