### 2021-04-08 vprusso/toqito#48 @vprusso
Hi @rurz,

Thank you for the kind words on the toqito software package (and apologies for my belated response to this thread)!

Based on your great comments, I've created the following pull request: https://github.com/vprusso/toqito/pull/51

> Minor typos in the documentation (It denotes that was a human who do the work)

I did my best to comb through the docs (tutorials, auto-generated docs from comments in functions, etc.) to try to eliminate some of the typos and spelling errors I came across. I'm sure there will inevitably be some that I've missed!

> In the paper a typo: tranpose -> transpose

Nice catch. Fixed in #50.

> I receive errors from the Lower bounding the quantum value of the CHSH game example tutorial

Right, this was also something found and fixed in #50. The tutorial examples were not updated when the rest of the code was updated, so that was my fault there.

> This last error repeats in the FFL game. How this errors can be solved to proceed to check what the tutorial claims?

Similar issue as to the one above--fixed in #50, and my fault that I did not update the tutorials.

Let me know if this satisfies your review, and thank you again for taking the time and making the effort to improve the toqito package!
