### 2021-04-12 57259dc
Updating getting_started markdown to RST format. Fixing small comments.
