### 2021-04-15 vprusso/toqito#48 @rurz
> Let me know if this satisfies your review, and thank you again for taking the time and making the effort to improve the toqito package!

Hi @vprusso, I'm sorry for the late response, it was holiday season at Mexico.

Great, I see a lot of work done in the #50. Let me recheck again, and I find something to improve, I will notify you. I think two or three days must be enough. Anticipating congrats for the amazing package you have here. 👍🏻
