### 2021-04-16 vprusso/toqito#48 @vprusso
A belated happy holiday season to you and yours, @rurz!

Do let me know if you stumble on anything that should be improved, and thank you yet again for your insightful and well-written set of initial comments!
