### 2021-04-17 vprusso/toqito#54 @jameshclrk
JOSS Review
Hello,

I like the package. I wish I has something like this when I first started learning about quantum information :smile:

I had a few comments:

### Functionality
- I had a few errors when trying out the library, relating to BLAS/LAPACK packages for the SCS solver in cvxpy. While this isn't a mistake in this package, it might be worthwhile to document the prerequisites for the SCS solver
- When trying the non-signalling values (Extended CHSH game) I get the following warning message (although I get the correct result.):
```
WARN: A->p (column pointers) not strictly increasing, column 48 empty
```

### Documentation
The docs are great. Love the detailed tutorials on games.
- Installation instructions could mention BLAS/LAPACK issues above
- An "Issue" template or guide would be useful in addition to the PR documentation.
- Some issues with math not showing up in the tutorial (https://toqito.readthedocs.io/en/latest/tutorials.nonlocal_games.html#example-lower-bounding-the-quantum-value-of-the-chsh-game)
Minor issues:
- Some code blocks have the `>>>` from the REPL and some code blocks don't. I would pick a format.
- Reference labels have inconsistent formats in docs (This is small and I think doesn't affect the review, but nice to fix later)
- Another comment for future documentation would be for tutorials in the channels.

### Paper
- It seems good to me just one comment: transpose has a typo - I checked the master branch and its still there

Overall, great project. Looking forward to what v2 will have ;)
