### 2021-04-18 8eb42ae
Responding to comments from jameshclrk for JOSS review.
