+++ toqito/state_props/entanglement_of_formation.py
@@ -53,7 +53,7 @@ def entanglement_of_formation(rho: np.ndarray, dim: Union[List[int], int] = None
     :param rho: A matrix or vector.
     :param dim: The default has both subsystems of equal dimension.
     :return: A value between 0 and 1 that corresponds to the
-             entanglement-of-formation of :code:`\rho`.
+             entanglement-of-formation of :code:`rho`.
     """
     dim_x, dim_y = rho.shape
     round_dim = int(np.round(np.sqrt(max(dim_x, dim_y))))
