+++ toqito/state_props/is_mutually_orthogonal.py
@@ -52,7 +52,8 @@ def is_mutually_orthogonal(vec_list: List[Union[np.ndarray, List[Union[float, An
         https://en.wikipedia.org/wiki/Orthogonality
 
     :param vec_list: The list of vectors to check.
-    :return: True if :code:`vec_list` are mutually orthogonal, and False otherwise.
+    :return: :code:`True` if :code:`vec_list` are mutually orthogonal, and
+             :code:`False` otherwise.
     """
     if len(vec_list) <= 1:
         raise ValueError("There must be at least two vectors provided as input.")
