### 2021-04-18 vprusso/toqito#54 @vprusso
Hi @jameshclrk,

First of all, thank you very much for taking the time to go over the fine-grained details of the toqito project. Your inputs will undoubtedly be incredibly important to improving the project, so thank you very much for that.

> I had a few errors when trying out the library, relating to BLAS/LAPACK packages for the SCS solver in cvxpy. While this isn't a mistake in this package, it might be worthwhile to document the prerequisites for the SCS solver

Great tip. I made sure to include some extra information and links in the "Getting Started" (https://toqito.readthedocs.io/en/latest/getting_started.html#) installation page.

> When trying the non-signaling values (Extended CHSH game) I get the following warning message (although I get the correct result.):
WARN: A->p (column pointers) not strictly increasing, column 48 empty

Right, while I am aware of this warning, it does not appear to impact the correctness of the returned value. I did however add in a clarifying note in the tutorial page for extended nonlocal games to notify the user. I've also placed an issue on the issues board (https://github.com/vprusso/toqito/issues/55) for the toqito project. I will continue to look into why the warnings may be popping up here.

> The docs are great. Love the detailed tutorials on games.

Thanks so much, that's wonderful to hear!

> Some code blocks have the >>> from the REPL and some code blocks don't. I would pick a format.

I made sure to add those `>>>` in for consistency. Thanks, and nice catch!

> Reference labels have inconsistent formats in docs (This is small and I think doesn't affect the review, but nice to fix later)

Right, this is something I may leave for a bit later--the reason being is that there appears to be something odd going on with having references from other documents that have the same reference tags (which led to the inconsistency in reference naming). I will add this to the issues board as a future fix as per your suggestion though!

> Another comment for future documentation would be for tutorials in the channels.

Yep! I am presently in the process of beefing up some of the existing docs in addition to adding some extra docs--specifically for channels. Good suggestion!

> It seems good to me just one comment: transpose has a typo - I checked the master branch and it's still there

Dang, not sure how that snuck back in there. Hopefully, it's fixed now!

> Overall, great project. Looking forward to what v2 will have ;)

Thanks very much, me too! I'm working on some new features that I'm excited to include in the upcoming version. Stay tuned, and thanks again for all your input!
