### 2021-04-19 vprusso/toqito#48 @rurz
> Do let me know if you stumble on anything that should be improved, and thank you yet again for your insightful and well-written set of initial comments!

Hi again, @vprusso. I have rechecked my initial statements of this review, and I see that the work done in #50 neatly solves the issues. Congrats for such effort. Currently, I have no complains about installation, functionality and documentation of the package, neither the companion paper. I check this review as done, and you can close the thread. I will complete my review on JOSS repo.

By the way, I nice addition would be a translation to Spanish of the documentation. I would be pleasure to contribute in that front.
