### 2021-04-20 vprusso/toqito#54 @jameshclrk
Cool! Looks great. I'll close this issue now, all seems good.

> Thanks very much, me too! I'm working on some new features that I'm excited to include in the upcoming version. Stay tuned, and thanks again for all your input!
No problem. Thanks for building this package.
