### 2021-04-24 a051ff7
Fixing bugs in permute_systems
