+++ tests/test_perms/test_permute_systems.py
@@ -73,11 +73,99 @@ def permute_systems_test_2_3_1():
     np.testing.assert_equal(np.all(bool_mat), True)
 
 
+def test_permute_systems_16_by_16():
+    """Permute systems of dimension 16-by-16."""
+    rho = np.array(
+        [
+            [1, 2, 33, 34, 65, 66, 97, 98, 9, 10, 41, 42, 73, 74, 105, 106],
+            [17, 18, 49, 50, 81, 82, 113, 114, 25, 26, 57, 58, 89, 90, 121, 122],
+            [3, 4, 35, 36, 67, 68, 99, 100, 11, 12, 43, 44, 75, 76, 107, 108],
+            [19, 20, 51, 52, 83, 84, 115, 116, 27, 28, 59, 60, 91, 92, 123, 124],
+            [5, 6, 37, 38, 69, 70, 101, 102, 13, 14, 45, 46, 77, 78, 109, 110],
+            [21, 22, 53, 54, 85, 86, 117, 118, 29, 30, 61, 62, 93, 94, 125, 126],
+            [7, 8, 39, 40, 71, 72, 103, 104, 15, 16, 47, 48, 79, 80, 111, 112],
+            [23, 24, 55, 56, 87, 88, 119, 120, 31, 32, 63, 64, 95, 96, 127, 128],
+            [129, 130, 161, 162, 193, 194, 225, 226, 137, 138, 169, 170, 201, 202, 233, 234],
+            [145, 146, 177, 178, 209, 210, 241, 242, 153, 154, 185, 186, 217, 218, 249, 250],
+            [131, 132, 163, 164, 195, 196, 227, 228, 139, 140, 171, 172, 203, 204, 235, 236],
+            [147, 148, 179, 180, 211, 212, 243, 244, 155, 156, 187, 188, 219, 220, 251, 252],
+            [133, 134, 165, 166, 197, 198, 229, 230, 141, 142, 173, 174, 205, 206, 237, 238],
+            [149, 150, 181, 182, 213, 214, 245, 246, 157, 158, 189, 190, 221, 222, 253, 254],
+            [135, 136, 167, 168, 199, 200, 231, 232, 143, 144, 175, 176, 207, 208, 239, 240],
+            [151, 152, 183, 184, 215, 216, 247, 248, 159, 160, 191, 192, 223, 224, 255, 256],
+        ]
+    )
+    perm = [2, 3, 1, 4]
+    dim = [[2, 2, 2, 2], [2, 2, 2, 2]]
+
+    res = permute_systems(rho, perm, dim, True, False)
+    expected_res = np.array(
+        [
+            [1, 2, 33, 34, 65, 66, 97, 98, 9, 10, 41, 42, 73, 74, 105, 106],
+            [17, 18, 49, 50, 81, 82, 113, 114, 25, 26, 57, 58, 89, 90, 121, 122],
+            [129, 130, 161, 162, 193, 194, 225, 226, 137, 138, 169, 170, 201, 202, 233, 234],
+            [145, 146, 177, 178, 209, 210, 241, 242, 153, 154, 185, 186, 217, 218, 249, 250],
+            [3, 4, 35, 36, 67, 68, 99, 100, 11, 12, 43, 44, 75, 76, 107, 108],
+            [19, 20, 51, 52, 83, 84, 115, 116, 27, 28, 59, 60, 91, 92, 123, 124],
+            [131, 132, 163, 164, 195, 196, 227, 228, 139, 140, 171, 172, 203, 204, 235, 236],
+            [147, 148, 179, 180, 211, 212, 243, 244, 155, 156, 187, 188, 219, 220, 251, 252],
+            [5, 6, 37, 38, 69, 70, 101, 102, 13, 14, 45, 46, 77, 78, 109, 110],
[…]
