### 2021-04-25 #3082 @poulson
@vprusso: Would you mind fixing the PDF compilation issue. It looks like the 'affiliations' field is missing.
