### 2021-04-27 b8a56a1
Fixing issue in 'swap' function.
