+++ tests/test_perms/test_swap.py
@@ -4,14 +4,24 @@ import numpy as np
 from toqito.perms import swap
 
 
-def test_swap_matrix():
-    """Tests swap operation on matrix."""
-    test_mat = np.array([[1, 5, 9, 13], [2, 6, 10, 14], [3, 7, 11, 15], [4, 8, 12, 16]])
+def test_swap_matrix_4_by_4():
+    """Tests swap operation on a 4-by-4 matrix."""
+    test_mat = np.arange(1, 17).reshape(4, 4).transpose()
 
+    # Default argument (systems [1, 2]):
     expected_res = np.array([[1, 9, 5, 13], [3, 11, 7, 15], [2, 10, 6, 14], [4, 12, 8, 16]])
-
     res = swap(test_mat)
 
+    bool_mat = np.isclose(res, expected_res)
+    np.testing.assert_equal(np.all(bool_mat), True)
+
+    # Hard-coded argument (systems [1, 2]):
+    res = swap(test_mat, [1, 2], [2, 2])
+    bool_mat = np.isclose(res, expected_res)
+    np.testing.assert_equal(np.all(bool_mat), True)
+
+    # Hard-coded argument (systems [2, 1])--this should be identical to the prior two cases.
+    res = swap(test_mat, [2, 1], [2, 2])
     bool_mat = np.isclose(res, expected_res)
     np.testing.assert_equal(np.all(bool_mat), True)
 
@@ -41,24 +51,258 @@ def test_swap_int_dim():
 
 
 def test_swap_8_by_8():
-    """Test swap on an 8x8 matrix."""
-    test_mat = np.arange(64).reshape(8, 8).T
+    """Test swap on an 8-by-8 matrix."""
+    test_mat = np.arange(1, 65).reshape(8, 8).transpose()
 
-    expected_res = np.array(
+    # Swapping on systems "1" and "2":
+    res_1_2 = swap(test_mat, [1, 2], [2, 2, 2])
+    expected_res_1_2 = np.array(
         [
-            [0, 16, 8, 24, 32, 48, 40, 56],
-            [2, 18, 10, 26, 34, 50, 42, 58],
-            [1, 17, 9, 25, 33, 49, 41, 57],
-            [3, 19, 11, 27, 35, 51, 43, 59],
-            [4, 20, 12, 28, 36, 52, 44, 60],
-            [6, 22, 14, 30, 38, 54, 46, 62],
-            [5, 21, 13, 29, 37, 53, 45, 61],
-            [7, 23, 15, 31, 39, 55, 47, 63],
+            [1, 9, 33, 41, 17, 25, 49, 57],
+            [2, 10, 34, 42, 18, 26, 50, 58],
+            [5, 13, 37, 45, 21, 29, 53, 61],
+            [6, 14, 38, 46, 22, 30, 54, 62],
+            [3, 11, 35, 43, 19, 27, 51, 59],
+            [4, 12, 36, 44, 20, 28, 52, 60],
+            [7, 15, 39, 47, 23, 31, 55, 63],
+            [8, 16, 40, 48, 24, 32, 56, 64],
         ]
     )
-    res = swap(test_mat, [2, 3], [2, 2, 2])
+    bool_mat = np.isclose(res_1_2, expected_res_1_2)
+    np.testing.assert_equal(np.all(bool_mat), True)
 
-    bool_mat = np.isclose(res, expected_res)
+    # Swapping on systems "2" and "1" (should be same as swapping on "1" and "2"):
+    res_2_1 = swap(test_mat, [2, 1], [2, 2, 2])
[…]
