+++ toqito/nonlocal_games/nonlocal_game.py
@@ -294,11 +294,9 @@ class NonlocalGame:
 
                 # As the SDPs keep alternating, check if the winning probability
                 # becomes any higher. If so, replace with new best.
-                if best < lower_bound:
-                    best = lower_bound
+                best = max(best, lower_bound)
 
-            if best_lower_bound < best:
-                best_lower_bound = best
+            best_lower_bound = max(best, best_lower_bound)
 
         return best_lower_bound
 
