### 2021-05-01 9ee3a54
Fixing Hadamard matrix issue.
