### 2021-05-02 #3082 @vprusso
@poulson. Fixed in https://github.com/vprusso/toqito/commit/537ee6051aa5355bbb09e6e12115f1f1d6b43ebc. Thank you very much for the catch!
