### 2021-05-03 #3082 @poulson
Hi @vprusso -- it is looking good but I would consider shortening "applications pertaining to physics" to "physics applications" and potentially replace the comma after "namely" with a colon given the list that follows, and potentially also add a colon after "pertaining to".

With that said, these are minor.

Would you mind registering for a DOI on zenodo (if this version of the software does not already have one) and sharing it here?
