### 2021-05-05 0cb52bd
Updating paper with JOSS comments.
