### 2021-05-05 #3082 @vprusso
Hi @poulson. All great suggestions. Updated accordingly in https://github.com/vprusso/toqito/commit/0cb52bd625de25f8149903f7f52ade91b918ed1f

And yes, I do have a DOI on Zenodo which I can provide below:
10.5281/zenodo.3699578

Thanks again for your comments, and let me know if I can provide anything else. Thanks again!
