### 2021-05-07 #3082 @kthyng
Hi @vprusso! Can I verify what the current version of your software associated with this JOSS publication is? Also, I see you included a Zenodo archive above but given that it is from March 2020, I'm thinking the software has changed since then. Can you make a Zenodo archive for the software as it stands? This will be associated with your publication. Also, please alter the metadata of the Zenodo archive so that the author list and title exactly match the JOSS paper. Thanks.
