### 2021-05-08 648a616
Fixing references in paper for JOSS. Udpating DOI reference.
