+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{diamond2016cvxpy,
-  title={CVXPY: A Python-embedded modeling language for convex optimization},
+  title={{CVXPY}: A {P}ython-embedded modeling language for convex optimization},
   author={Diamond, Steven and Boyd, Stephen},
   journal={The Journal of Machine Learning Research},
   volume={17},
@@ -18,7 +18,7 @@
 }
 
 @article{johansson2013qutip,
-  title={QuTiP 2: A Python framework for the dynamics of open quantum systems},
+  title={{Q}u{T}i{P} 2: A {P}ython framework for the dynamics of open quantum systems},
   author={Johansson, J Robert and Nation, Paul D and Nori, Franco},
   journal={Computer Physics Communications},
   volume={184},
@@ -41,7 +41,7 @@
 }
 
 @article{steiger2018projectq,
-  title={ProjectQ: an open source software framework for quantum computing},
+  title={Project{Q}: an open source software framework for quantum computing},
   author={Steiger, Damian S and H{\"a}ner, Thomas and Troyer, Matthias},
   journal={Quantum},
   volume={2},
@@ -52,7 +52,7 @@
 }
 
 @article{gheorghiu2014quantum++,
-  title={Quantum++--A C++ 11 quantum computing library},
+  title={Quantum++--{A} {C}++ 11 quantum computing library},
   author={Gheorghiu, Vlad},
   journal={arXiv preprint arXiv:1412.4704},
   doi={10.1371/journal.pone.0208073},
