+++ docs/index.rst
@@ -20,9 +20,9 @@ that are often associated with computer science.
 .. image:: https://codecov.io/gh/vprusso/toqito/branch/master/graph/badge.svg?style=plastic
    :alt: Codecov
    :target: https://codecov.io/gh/vprusso/toqito
-.. image:: https://zenodo.org/badge/235493396.svg?style=plastic
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4743211.svg?style=plastic
    :alt: DOI
-   :target: https://zenodo.org/badge/latestdoi/235493396
+   :target: https://doi.org/10.5281/zenodo.4743211
 .. image:: https://img.shields.io/badge/Supported%20By-UNITARY%20FUND-brightgreen.svg?style=plastic
    :alt: Unitary Fund
    :target: http://unitary.fund
