### 2021-05-08 #3082 @vprusso
Hi @kthyng. Excellent points. Let me address those below:

> Can I verify what the current version of your software associated with this JOSS publication is?

Sure, the current version is `v1.0`

> Also, I see you included a Zenodo archive above but given that it is from March 2020, I'm thinking the software has changed since then. Can you make a Zenodo archive for the software as it stands? This will be associated with your publication

Very good point, I've updated that and have the new Zenodo archive provided here:
https://zenodo.org/record/4743211

> Also, please alter the metadata of the Zenodo archive so that the author list and title exactly match the JOSS paper. Thanks.

Done!

> Please check out the example paper and update your paper so that the parenthetical references are together in one set of parentheses instead of separated out one by one. Also, your references have lower case "python"s in them (and possibly other words). Please update your .bib file so that the references have proper capitalization. You can preserve capitalization with {} around characters, words, or phrases.

Done and updated in the following commit: https://github.com/vprusso/toqito/commit/648a6164f826760ee775319af8a861e6480f1f1b

Thank you again for all the attention to detail. Do let me know if there is anything else to address, and thank you again!
