### 2021-05-09 a4b74b6
Fixing bug in partial trace. Adding more tests to partial trace and partial transpose.
