+++ tests/test_channels/test_partial_trace.py
@@ -74,67 +74,144 @@ def test_partial_trace_sys_int_dim_int_2():
     np.testing.assert_equal(np.all(bool_mat), True)
 
 
+def test_partial_trace_4_by_4():
+    """Test for 4-by-4 matrix."""
+    test_input_mat = np.arange(1, 17).reshape(4, 4)
+
+    pt_1 = partial_trace(test_input_mat, [1], [2, 2])
+    pt_2 = partial_trace(test_input_mat, [2], [2, 2])
+
+    expected_pt_1 = np.array([[12, 14], [20, 22]])
+    bool_mat = np.isclose(expected_pt_1, pt_1)
+    np.testing.assert_equal(np.all(bool_mat), True)
+
+    expected_pt_2 = np.array([[7, 11], [23, 27]])
+    bool_mat = np.isclose(expected_pt_2, pt_2)
+    np.testing.assert_equal(np.all(bool_mat), True)
+
+
 def test_partial_trace_8_by_8():
     """Test for 8-by-8 matrix."""
     test_input_mat = np.arange(1, 65).reshape(8, 8)
-    res = partial_trace(test_input_mat, [1, 2], [2, 2, 2])
 
-    expected_res = np.array([[112, 116], [144, 148]])
-
-    bool_mat = np.isclose(expected_res, res)
-    np.testing.assert_equal(np.all(bool_mat), True)
-
-
-def test_partial_trace_8_by_8_2():
-    """Test for 8-by-8 matrix."""
-    test_input_mat = np.arange(1, 65).reshape(8, 8)
-    res = partial_trace(test_input_mat, [1], [2, 2, 2])
-
-    expected_res = np.array(
+    # Trace out first subsystem:
+    pt_1 = partial_trace(test_input_mat, [1], [2, 2, 2])
+    expected_pt_1 = np.array(
         [[38, 40, 42, 44], [54, 56, 58, 60], [70, 72, 74, 76], [86, 88, 90, 92]]
     )
-
-    bool_mat = np.isclose(expected_res, res)
+    bool_mat = np.isclose(expected_pt_1, pt_1)
     np.testing.assert_equal(np.all(bool_mat), True)
 
+    # Trace out second subsystem:
+    pt_2 = partial_trace(test_input_mat, [2], [2, 2, 2])
+    expected_pt_2 = np.array(
+        [[20, 22, 28, 30], [36, 38, 44, 46], [84, 86, 92, 94], [100, 102, 108, 110]]
+    )
+    bool_mat = np.isclose(expected_pt_2, pt_2)
+    np.testing.assert_equal(np.all(bool_mat), True)
 
-def test_partial_trace_8_by_8_3():
-    """Test for 8-by-8 matrix."""
-    test_input_mat = np.arange(1, 65).reshape(8, 8)
-    res = partial_trace(test_input_mat, 3, [2, 2, 2])
-
-    expected_res = np.array(
+    # Trace out third subsystem:
+    pt_3 = partial_trace(test_input_mat, [3], [2, 2, 2])
+    expected_pt_3 = np.array(
         [[11, 15, 19, 23], [43, 47, 51, 55], [75, 79, 83, 87], [107, 111, 115, 119]]
     )
+    bool_mat = np.isclose(expected_pt_3, pt_3)
+    np.testing.assert_equal(np.all(bool_mat), True)
 
-    bool_mat = np.isclose(expected_res, res)
+    # Trace out first and second subsystem:
+    pt_1_2 = partial_trace(test_input_mat, [1, 2], [2, 2, 2])
+    expected_pt_1_2 = np.array([[112, 116], [144, 148]])
+    bool_mat = np.isclose(expected_pt_1_2, pt_1_2)
[…]
