+++ tests/test_channels/test_partial_transpose.py
@@ -103,8 +103,227 @@ def test_partial_transpose_norm_diff():
     np.testing.assert_equal(np.isclose(res, expected_res), True)
 
 
-def test_partial_transpose_16_by_16():
-    """Partial transpose on a 16-by-16 matrix."""
+def test_partial_transpose_8_by_8_subsystems_2_2_2():
+    """Partial transpose on a 8-by-8 matrix on 2 x 2 x 2 subsystems."""
+    test_input_mat = np.arange(1, 65).reshape(8, 8)
+
+    # Partial transpose on first subsystem:
+    pt_1 = partial_transpose(test_input_mat, [1], [2, 2, 2])
+    expected_pt_1 = np.array(
+        [
+            [1, 2, 3, 4, 33, 34, 35, 36],
+            [9, 10, 11, 12, 41, 42, 43, 44],
+            [17, 18, 19, 20, 49, 50, 51, 52],
+            [25, 26, 27, 28, 57, 58, 59, 60],
+            [5, 6, 7, 8, 37, 38, 39, 40],
+            [13, 14, 15, 16, 45, 46, 47, 48],
+            [21, 22, 23, 24, 53, 54, 55, 56],
+            [29, 30, 31, 32, 61, 62, 63, 64],
+        ]
+    )
+    bool_mat = np.isclose(expected_pt_1, pt_1)
+    np.testing.assert_equal(np.all(bool_mat), True)
+
+    # Partial transpose on second subsystem:
+    pt_2 = partial_transpose(test_input_mat, [2], [2, 2, 2])
+    expected_pt_2 = np.array(
+        [
+            [1, 2, 17, 18, 5, 6, 21, 22],
+            [9, 10, 25, 26, 13, 14, 29, 30],
+            [3, 4, 19, 20, 7, 8, 23, 24],
+            [11, 12, 27, 28, 15, 16, 31, 32],
+            [33, 34, 49, 50, 37, 38, 53, 54],
+            [41, 42, 57, 58, 45, 46, 61, 62],
+            [35, 36, 51, 52, 39, 40, 55, 56],
+            [43, 44, 59, 60, 47, 48, 63, 64],
+        ]
+    )
+    bool_mat = np.isclose(expected_pt_2, pt_2)
+    np.testing.assert_equal(np.all(bool_mat), True)
+
+    # Partial transpose on third subsystem:
+    pt_3 = partial_transpose(test_input_mat, [3], [2, 2, 2])
+    expected_pt_3 = np.array(
+        [
+            [1, 9, 3, 11, 5, 13, 7, 15],
+            [2, 10, 4, 12, 6, 14, 8, 16],
+            [17, 25, 19, 27, 21, 29, 23, 31],
+            [18, 26, 20, 28, 22, 30, 24, 32],
+            [33, 41, 35, 43, 37, 45, 39, 47],
+            [34, 42, 36, 44, 38, 46, 40, 48],
+            [49, 57, 51, 59, 53, 61, 55, 63],
+            [50, 58, 52, 60, 54, 62, 56, 64],
+        ]
+    )
+    bool_mat = np.isclose(expected_pt_3, pt_3)
+    np.testing.assert_equal(np.all(bool_mat), True)
+
+    # Partial transpose on first and second subsystem:
+    pt_1_2 = partial_transpose(test_input_mat, [1, 2], [2, 2, 2])
+    expected_pt_1_2 = np.array(
+        [
+            [1, 2, 17, 18, 33, 34, 49, 50],
+            [9, 10, 25, 26, 41, 42, 57, 58],
+            [3, 4, 19, 20, 35, 36, 51, 52],
[…]
